refactor(launcher): extract saga handlers into reducer/saga.rs (task #182 PR-D)
Extracts the 2 saga-related handlers from reducer/mod.rs. Both are
pure pass-through (state untouched; just translates the wire command
into the typed event for the saga coordinator's bus loop).

Extracted:
- handle_report_panes_reaped (~33 lines incl. doc)
- handle_report_saga_action_failed (~17 lines incl. doc)

Both marked pub(super). Dispatcher arms updated to saga::handle_*.

Sizes:
- reducer/mod.rs: 815 → 765 lines (-50)
- reducer/saga.rs: NEW 66 lines

154 launcher tests pass.
